Use common macro.
